requirements.txt
pytest-playwright
python-box
pyyaml
unstructured